fixed a memory leak
i was creating new tiles and copying them into the tile array, creating a memory leak where the pointers to these new tiles were gone but the tiles existed, and duplicates were stored in the tile array. this caused slow starting times as it had to fill up the array with duplicate tiles that were on the heap.